// FontAdmin_Tech/src/app/Models/nhan-vien.ts
export interface INhanVien {
    id?: number
    roleId?: number
    email?: string
    passWord?: string
    avatar?: string
    hoTen?: string
    soDienThoai?: number
    diaChi?: string
    gioiTinh?: string
    ngaySinh?: Date
    ngayVaoLam?: Date
    chucVuId?: number
    trangThai?: boolean
}